Refactor normalization functions to static methods
